refactor: empty app-test 400-line allowlist by densifying remaining suites
Extract fixtures and concern-split sibling suites so every app/test file
is at or under 400 lines, closing Slice D AC8 without new partN shards.
