+++ app/test/app_event_handler_test.dart
@@ -8,8 +8,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:colonizethis_app/config/route_paths.dart';
 import 'package:colonizethis_app/core/services/app_event_handler/app_event_handler.dart';
 import 'package:colonizethis_app/core/utils/state_toggle_notifier.dart';
-import 'package:colonizethis_app/features/game/widgets/combat/combat_mode_choice_dialog.dart';
-import 'package:colonizethis_app/features/game/widgets/combat/quick_battle_result_dialog.dart';
 import 'package:colonizethis_app/features/game/widgets/panels/pause_menu_panel.dart';
 import 'package:colonizethis_app/providers/turn_resolution_blocking_provider.dart';
 
@@ -349,73 +347,5 @@ void main() {
         expect(find.text('dialog:after_close'), findsOneWidget);
       },
     );
-
-    testWidgets(
-      'CombatModeChoiceDialog opened via OpenDialog emits CombatModeChosenEvent',
-      (tester) async {
-        handler = AppEventHandler(
-          bus: bus,
-          navigatorKey: navKey,
-          dialogBuilders: {
-            'combat_mode_choice': (ctx, params) => CombatModeChoiceDialog(
-              bus: bus,
-              provinceName: params?['provinceName'] as String? ?? '',
-              isCapitalSiege: params?['isCapitalSiege'] as bool? ?? false,
-            ),
-          },
-        );
-        handler.bind();
-        CombatModeChosenEvent? chosen;
-        bus.on<CombatModeChosenEvent>().listen((e) => chosen = e);
-        await pumpAppEventHandlerEmitButton(
-          tester,
-          navigatorKey: navKey,
-          label: 'open',
-          onPressed: () => bus.emit(
-            const OpenDialogEvent('combat_mode_choice', {
-              'provinceName': 'TestProv',
-              'isCapitalSiege': false,
-            }),
-          ),
-        );
-        await tapAppEventHandlerLabel(tester, 'open');
-        await tapAppEventHandlerLabel(tester, 'Quick Battle');
-        expect(chosen?.mode, CombatMode.quickBattle);
-        expect(find.text('Combat at TestProv'), findsNothing);
-      },
-    );
-
-    testWidgets('OpenDialogEvent quick_battle_result shows dialog', (
-      tester,
-    ) async {
-      handler = AppEventHandler(
-        bus: bus,
-        navigatorKey: navKey,
-        dialogBuilders: {
-          'quick_battle_result': (ctx, params) => QuickBattleResultDialog(
-            result: params!['result'] as QuickBattleResult,
-            attackerName: 'A',
-            defenderName: 'D',
-          ),
-        },
-      );
-      handler.bind();
-      final qb = QuickBattleResult(
-        winner: QuickBattleWinner.attacker,
-        attackerCasualties: const [],
-        defenderCasualties: const [],
-        provinceFlips: false,
-      );
-      await pumpAppEventHandlerEmitButton(
-        tester,
-        navigatorKey: navKey,
-        label: 'open',
-        onPressed: () =>
-            bus.emit(OpenDialogEvent('quick_battle_result', {'result': qb})),
-      );
-      await tapAppEventHandlerLabel(tester, 'open');
-      expect(find.textContaining('Battle Result'), findsOneWidget);
-      expect(find.textContaining('A wins'), findsOneWidget);
-    });
   });
 }
