+++ app/test/trade_screen_market_tab_treasury_bid_cap_test.dart
@@ -1,52 +1,14 @@
 // Widget tests for the Trade Market tab treasury bid cap (Refs #3093 —
-// SPEC/ui/trade-screen.md § Market tab — treasury bid cap,
-// SPEC/game/world-market.md § Treasury budget for bids.
-//     treasury (conservative clamp per SPEC).
+// SPEC/ui/trade-screen.md § Market tab; SPEC/game/world-market.md).
 import 'package:colonizethis_app/features/game/screens/trade/trade_screen.dart';
-import 'package:colonizethis_app/providers/games_provider.dart';
-import 'package:colonizethis_app/providers/treasury_summary_provider.dart';
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_riverpod/misc.dart' show Override;
 import 'package:flutter_test/flutter_test.dart';
 
+import 'trade_screen_market_tab_treasury_bid_cap_support.dart';
 import 'trade_screen_test_support.dart';
-
-const String _humanPlayerId = kTradeTestHumanPlayerId;
-
-const CommodityId _timber = 'timber';
-const CommodityId _iron = 'iron';
-const CommodityId _lumber = 'lumber';
-
-Override _treasurySummaryOverride(Game game, int nonBidProjectedDelta) {
-  final Player player = game.players.first;
-  return treasurySummaryProvider.overrideWith((ref) {
-    final Orders orders = ref.watch(currentOrdersProvider);
-    final int bidSpend = stagedBidTotalSpendByPlayer(
-      orders: orders,
-      playerId: _humanPlayerId,
-      game: game,
-      resourceRules: ResourceRules.defaultRules,
-    );
-    return TreasurySummary(
-      treasury: player.treasury,
-      projectedDelta: nonBidProjectedDelta - bidSpend,
-    );
-  });
-}
-
-TradeOrder? _stagedOrder(ProviderContainer container, CommodityId commodityId) {
-  final Orders orders = container.read(currentOrdersProvider);
-  final List<TradeOrder>? list = orders.tradeOrdersByPlayerId[_humanPlayerId];
-  if (list == null) return null;
-  for (final TradeOrder o in list) {
-    if (o.commodityId == commodityId) return o;
-  }
-  return null;
-}
 
 void main() {
   suppressLogsForTests();
@@ -58,13 +20,18 @@ void main() {
       (tester) async {
         final ProviderContainer container = await pumpTradeScreenWithContainer(
           tester,
-          game: buildTradeTestGame(treasury: 100, prices: const {_timber: 30}),
+          game: buildTradeTestGame(
+            treasury: 100,
+            prices: const {kTreasuryBidTimber: 30},
+          ),
         );
 
-        await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_timber)));
-        await tester.pump();
+        await tapTreasuryMarketBid(tester, kTreasuryBidTimber);
 
-        final TradeOrder? staged = _stagedOrder(container, _timber);
+        final TradeOrder? staged = stagedTreasuryBidOrder(
+          container,
+          kTreasuryBidTimber,
+        );
         expect(staged, isNotNull);
         expect(staged!.type, TradeOrderType.bid);
         expect(
@@ -83,25 +50,23 @@ void main() {
     ) async {
       final ProviderContainer container = await pumpTradeScreenWithContainer(
         tester,
-        game: buildTradeTestGame(treasury: 100, prices: const {_timber: 30}),
-        initialOrders: Orders(
-          tradeOrdersByPlayerId: {
-            _humanPlayerId: [
-              TradeOrder(
-                commodityId: _timber,
-                type: TradeOrderType.bid,
-                quantity: 3,
-                priority: 1,
-              ),
-            ],
-          },
+        game: buildTradeTestGame(
+          treasury: 100,
+          prices: const {kTreasuryBidTimber: 30},
+        ),
+        initialOrders: stagedTreasuryTradeOrders(
+          commodityId: kTreasuryBidTimber,
+          type: TradeOrderType.bid,
+          quantity: 3,
         ),
       );
 
-      await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowIncrementKey(_timber)));
-      await tester.pump();
+      await tapTreasuryMarketIncrement(tester, kTreasuryBidTimber);
 
-      final TradeOrder? staged = _stagedOrder(container, _timber);
+      final TradeOrder? staged = stagedTreasuryBidOrder(
+        container,
+        kTreasuryBidTimber,
+      );
       expect(
         staged?.quantity,
         3,
@@ -120,33 +85,28 @@ void main() {
           tester,
           game: buildTradeTestGame(
             treasury: 100,
-            prices: const {_timber: 30, _iron: 80},
+            prices: const {kTreasuryBidTimber: 30, kTreasuryBidIron: 80},
           ),
-          initialOrders: Orders(
-            tradeOrdersByPlayerId: {
-              _humanPlayerId: [
-                TradeOrder(
-                  commodityId: _timber,
-                  type: TradeOrderType.bid,
-                  quantity: 3,
-                  priority: 1,
-                ),
-              ],
-            },
+          initialOrders: stagedTreasuryTradeOrders(
+            commodityId: kTreasuryBidTimber,
+            type: TradeOrderType.bid,
+            quantity: 3,
           ),
         );
 
-        await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_iron)));
-        await tester.pump();
+        await tapTreasuryMarketBid(tester, kTreasuryBidIron);
 
         expect(
-          _stagedOrder(container, _iron),
+          stagedTreasuryBidOrder(container, kTreasuryBidIron),
           isNull,
           reason:
               'Refs #3093 — treasury headroom 10 cannot cover iron price 80; '
               'toggle must be a silent no-op.',
         );
-        expect(_stagedOrder(container, _timber)?.quantity, 3);
+        expect(
+          stagedTreasuryBidOrder(container, kTreasuryBidTimber)?.quantity,
+          3,
+        );
       },
     );
 
@@ -157,27 +117,22 @@ void main() {
         tester,
         game: buildTradeTestGame(
           treasury: 50,
-          prices: const {_timber: 30},
-          stockpile: const {_timber: 10},
+          prices: const {kTreasuryBidTimber: 30},
+          stockpile: const {kTreasuryBidTimber: 10},
         ),
-        initialOrders: Orders(
-          tradeOrdersByPlayerId: {
-            _humanPlayerId: [
-              TradeOrder(
-                commodityId: _timber,
-                type: TradeOrderType.offer,
-                quantity: 4,
-                priority: 1,
-              ),
-            ],
-          },
+        initialOrders: stagedTreasuryTradeOrders(
+          commodityId: kTreasuryBidTimber,
+          type: TradeOrderType.offer,
+          quantity: 4,
         ),
       );
 
-      await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_timber)));
-      await tester.pump();
+      await tapTreasuryMarketBid(tester, kTreasuryBidTimber);
 
-      final TradeOrder? staged = _stagedOrder(container, _timber);
+      final TradeOrder? staged = stagedTreasuryBidOrder(
+        container,
+        kTreasuryBidTimber,
+      );
       expect(staged, isNotNull);
       expect(staged!.type, TradeOrderType.bid);
       expect(
@@ -202,10 +157,12 @@ void main() {
           ),
         );
 
-        await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_lumber)));
-        await tester.pump();
+        await tapTreasuryMarketBid(tester, kTreasuryBidLumber);
 
-        final TradeOrder? staged = _stagedOrder(container, _lumber);
+        final TradeOrder? staged = stagedTreasuryBidOrder(
+          container,
+          kTreasuryBidLumber,
+        );
         expect(
           staged,
           isNotNull,
@@ -226,25 +183,23 @@ void main() {
         'the treasury cap)', (tester) async {
       final ProviderContainer container = await pumpTradeScreenWithContainer(
         tester,
-        game: buildTradeTestGame(treasury: 100, prices: const {_timber: 30}),
-        initialOrders: Orders(
-          tradeOrdersByPlayerId: {
-            _humanPlayerId: [
-              TradeOrder(
-                commodityId: _timber,
-                type: TradeOrderType.bid,
-                quantity: 2,
-                priority: 1,
-              ),
-            ],
-          },
+        game: buildTradeTestGame(
+          treasury: 100,
+          prices: const {kTreasuryBidTimber: 30},
+        ),
+        initialOrders: stagedTreasuryTradeOrders(
+          commodityId: kTreasuryBidTimber,
+          type: TradeOrderType.bid,
+          quantity: 2,
         ),
       );
 
-      await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowDecrementKey(_timber)));
-      await tester.pump();
+      await tapTreasuryMarketDecrement(tester, kTreasuryBidTimber);
 
-      expect(_stagedOrder(container, _timber)?.quantity, 1);
+      expect(
+        stagedTreasuryBidOrder(container, kTreasuryBidTimber)?.quantity,
+        1,
+      );
     });
 
     testWidgets(
@@ -261,18 +216,20 @@ void main() {
         // commitments).
         final Game game = buildTradeTestGame(
           treasury: 100,
-          prices: const {_timber: 30},
+          prices: const {kTreasuryBidTimber: 30},
         );
         final ProviderContainer container = await pumpTradeScreenWithContainer(
           tester,
           game: game,
-          extraOverrides: <Override>[_treasurySummaryOverride(game, -40)],
+          extraOverrides: <Override>[treasuryBidSummaryOverride(game, -40)],
         );
 
-        await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_timber)));
-        await tester.pump();
+        await tapTreasuryMarketBid(tester, kTreasuryBidTimber);
 
-        TradeOrder? staged = _stagedOrder(container, _timber);
+        TradeOrder? staged = stagedTreasuryBidOrder(
+          container,
+          kTreasuryBidTimber,
+        );
         expect(staged?.type, TradeOrderType.bid);
         expect(
           staged?.quantity,
@@ -280,22 +237,16 @@ void main() {
           reason: 'Budget 60 / price 30 = 2 headroom; default qty 1 fits.',
         );
 
-        await tester.tap(
-          find.byKey(TradeScreenMarketKeys.marketRowIncrementKey(_timber)),
-        );
-        await tester.pump();
-        staged = _stagedOrder(container, _timber);
+        await tapTreasuryMarketIncrement(tester, kTreasuryBidTimber);
+        staged = stagedTreasuryBidOrder(container, kTreasuryBidTimber);
         expect(
           staged?.quantity,
           2,
           reason: 'Spend grows to 60; still inside the 60 budget.',
         );
 
-        await tester.tap(
-          find.byKey(TradeScreenMarketKeys.marketRowIncrementKey(_timber)),
-        );
-        await tester.pump();
-        staged = _stagedOrder(container, _timber);
+        await tapTreasuryMarketIncrement(tester, kTreasuryBidTimber);
+        staged = stagedTreasuryBidOrder(container, kTreasuryBidTimber);
         expect(
           staged?.quantity,
           2,
@@ -315,19 +266,18 @@ void main() {
         // helper clamps the bid budget at 0. Even default qty 1 cannot land.
         final Game game = buildTradeTestGame(
           treasury: 50,
-          prices: const {_timber: 30},
+          prices: const {kTreasuryBidTimber: 30},
         );
         final ProviderContainer container = await pumpTradeScreenWithContainer(
           tester,
           game: game,
-          extraOverrides: <Override>[_treasurySummaryOverride(game, -60)],
+          extraOverrides: <Override>[treasuryBidSummaryOverride(game, -60)],
         );
 
-        await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_timber)));
-        await tester.pump();
+        await tapTreasuryMarketBid(tester, kTreasuryBidTimber);
 
         expect(
-          _stagedOrder(container, _timber),
+          stagedTreasuryBidOrder(container, kTreasuryBidTimber),
           isNull,
           reason:
               'Refs #3093 — pending non-bid deficit (60) exceeds raw treasury '
@@ -346,18 +296,20 @@ void main() {
         // (conservative clamp per SPEC § Treasury budget for bids).
         final Game game = buildTradeTestGame(
           treasury: 100,
-          prices: const {_iron: 80},
+          prices: const {kTreasuryBidIron: 80},
         );
         final ProviderContainer container = await pumpTradeScreenWithContainer(
           tester,
           game: game,
-          extraOverrides: <Override>[_treasurySummaryOverride(game, 25)],
+          extraOverrides: <Override>[treasuryBidSummaryOverride(game, 25)],
         );
 
-        await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_iron)));
-        await tester.pump();
+        await tapTreasuryMarketBid(tester, kTreasuryBidIron);
 
-        final TradeOrder? staged = _stagedOrder(container, _iron);
+        final TradeOrder? staged = stagedTreasuryBidOrder(
+          container,
+          kTreasuryBidIron,
+        );
         expect(staged?.type, TradeOrderType.bid);
         expect(
           staged?.quantity,
@@ -381,32 +333,22 @@ void main() {
         // treasury 100. Spend 60 + 30 = 90 ≤ 100, so the increment lands.
         final Game game = buildTradeTestGame(
           treasury: 100,
-          prices: const {_timber: 30},
+          prices: const {kTreasuryBidTimber: 30},
         );
         final ProviderContainer container = await pumpTradeScreenWithContainer(
           tester,
           game: game,
-          initialOrders: Orders(
-            tradeOrdersByPlayerId: {
-              _humanPlayerId: [
-                TradeOrder(
-                  commodityId: _timber,
-                  type: TradeOrderType.bid,
-                  quantity: 2,
-                  priority: 1,
-                ),
-              ],
-            },
+          initialOrders: stagedTreasuryTradeOrders(
+            commodityId: kTreasuryBidTimber,
+            type: TradeOrderType.bid,
+            quantity: 2,
           ),
-          extraOverrides: <Override>[_treasurySummaryOverride(game, 10)],
+          extraOverrides: <Override>[treasuryBidSummaryOverride(game, 10)],
         );
 
-        await tester.tap(
-          find.byKey(TradeScreenMarketKeys.marketRowIncrementKey(_timber)),
-        );
-        await tester.pump();
+        await tapTreasuryMarketIncrement(tester, kTreasuryBidTimber);
         expect(
-          _stagedOrder(container, _timber)?.quantity,
+          stagedTreasuryBidOrder(container, kTreasuryBidTimber)?.quantity,
           3,
           reason:
               'Reconstructed non-bid delta is +10 (net income); budget '
@@ -426,37 +368,32 @@ void main() {
         // would need 80 of remaining 30 headroom.
         final Game game = buildTradeTestGame(
           treasury: 100,
-          prices: const {_timber: 30, _iron: 80},
+          prices: const {kTreasuryBidTimber: 30, kTreasuryBidIron: 80},
         );
         final ProviderContainer container = await pumpTradeScreenWithContainer(
           tester,
           game: game,
-          initialOrders: Orders(
-            tradeOrdersByPlayerId: {
-              _humanPlayerId: [
-                TradeOrder(
-                  commodityId: _timber,
-                  type: TradeOrderType.bid,
-                  quantity: 1,
-                  priority: 1,
-                ),
-              ],
-            },
+          initialOrders: stagedTreasuryTradeOrders(
+            commodityId: kTreasuryBidTimber,
+            type: TradeOrderType.bid,
+            quantity: 1,
           ),
-          extraOverrides: <Override>[_treasurySummaryOverride(game, -40)],
+          extraOverrides: <Override>[treasuryBidSummaryOverride(game, -40)],
         );
 
-        await tester.tap(find.byKey(TradeScreenMarketKeys.marketRowBidChipKey(_iron)));
-        await tester.pump();
+        await tapTreasuryMarketBid(tester, kTreasuryBidIron);
 
         expect(
-          _stagedOrder(container, _iron),
+          stagedTreasuryBidOrder(container, kTreasuryBidIron),
           isNull,
           reason:
               'Bid budget 60 − existing timber spend 30 leaves 30 headroom; '
               'iron price 80 > 30 → toggle is silent no-op.',
         );
-        expect(_stagedOrder(container, _timber)?.quantity, 1);
+        expect(
+          stagedTreasuryBidOrder(container, kTreasuryBidTimber)?.quantity,
+          1,
+        );
       },
     );
   });
