+++ app/test/units_panel_shared_widgets_test_support.dart
@@ -1,7 +1,9 @@
 // Pump helpers for units-panel shared-widget tests (Refs #4352).
 
+import 'package:colonizethis_app/features/game/widgets/units/shared/region_section_header.dart';
 import 'package:colonizethis_app/features/game/widgets/units/shared/units_entity_action_row.dart';
 import 'package:colonizethis_app/features/game/widgets/units/shared/units_panel_shell.dart';
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -50,4 +52,30 @@ Future<void> pumpUnitsEntityActionRow(
       ),
     ),
   );
+}
+
+UnitsEntityAction unitsEntityMoveAction({
+  VoidCallback? onPressed,
+  bool enabled = true,
+}) {
+  return UnitsEntityAction(
+    tooltip: 'Move',
+    icon: Icons.route,
+    label: 'Move',
+    onPressed: enabled ? (onPressed ?? () {}) : null,
+  );
+}
+
+void expectRegionHeaderLeftBar(WidgetTester tester) {
+  final Border border = unitsSharedHeaderBorder(tester, RegionSectionHeader);
+  expect(border.left.width, RegionSectionHeader.leftBarWidth);
+  expect(border.left.color, EditorialMonoclePalette.accentDim);
+  expect(border.bottom, BorderSide.none);
+}
+
+void expectRegionHeaderBottomBorder(WidgetTester tester) {
+  final Border border = unitsSharedHeaderBorder(tester, RegionSectionHeader);
+  expect(border.bottom.width, RegionSectionHeader.bottomBorderWidth);
+  expect(border.bottom.color, EditorialMonoclePalette.border);
+  expect(border.left, BorderSide.none);
 }
